backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0,<0.115.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=1.10.17,<2.0.0",
    "sqlalchemy>=2.0.36",
    "psycopg2-binary==2.9.11",
    "alembic==1.13.1",
    "apscheduler==3.10.4",
    "httpx==0.26.0",
    "aiohttp==3.9.1",
    "python-dotenv==1.0.0",
    "python-multipart>=0.0.9",
    "email-validator==2.1.0",
    "cryptography==42.0.0",
    "python-json-logger==2.0.7",
    "structlog==24.1.0",
    "deepdiff==8.6.1",
    "snowflake-connector-python>=4.0.0",
    "redis>=5.0.0",
    "celery[redis]>=5.4.0,<6.0.0",
    "psutil>=7.2.2",
    "duckdb>=1.4.4",
    "pyarrow>=23.0.0",
]

[dependency-groups]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "black==23.12.1",
    "flake8==7.0.0",
    "mypy==1.8.0",
    "isort==5.13.2",
    "faker==22.0.0",
]

[tool.uv]
native-tls = true
